+++ packages/infra/src/tools/index.ts
@@ -1,3 +1,4 @@
 export * from "./bash.js";
 export * from "./browser.js";
 export * from "./git.js";
+export * from "./github.js";